Add OpenSSHShellSocket.
This socket needs to be used as a Custom Socket:

> $sock = new OpenSSHShellSocket('Host', 'User', 'Pass');
> $dev = new CiscoSwitch('', '', '', $sock);
> $dev->connect();

This isn't as well tested as the normal sockets, and may not behave
in all situations, but allows for connecting via JumpHosts for example:

> $sock = new OpenSSHShellSocket('Host', 'User', 'Pass');
> $sock->setParams('-o ProxyCommand="ssh jumphost -W Host:22"');
> $dev = new CiscoSwitch('', '', '', $sock);
> $dev->connect();

This also honors most aspects of .ssh/config, except for ensuring that
we do not make use of any ControlMaster sockets as these can change the
login procedure and cause the session to hang if this session ends up as
the ControlMaster and another session is opened. We also disable host-key
checking as this can also interfere with the login process.

We rely on `sshpass` to handle actually logging in and dealing with sending
the password.